# swagger.json
{
    "swagger":"2.0",
    "info": {
        "version": "1.0.0",
        "title": "Book store API",
        "description": "Rest API for Book Store"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/api/v1",
    "tags": [
        {
            "name": "book store",
            "description": "API for Book Store"
        }
    ],
    "schemes": ["http"],
    "consumes":["application/json"],
    "produces":["application/json"],
    "paths": {
        "/":{
            "get": {
                "tags": ["Books"],
                "summary": "Receive data",
                "produces":["application/json"],
                "responses": {
                    "200":{
                        "description": "Found",
                        "schema":{
                            
                        }
                      
                    }
                }
            }
        },
        "/view/{id}":{
            "get": {
                "tags": ["Obter livro"],
                "description": "Obtem o livro com o id associado",
                "produces":["application/json"],
                "parameters": [
                    {"in": "path",
                    "name": "id",
                    "required": true
                    }
                ],
                "responses": {
                    "200":{
                        "description": "Livro com o id passado no path"

                    },
                    "404":{
                        "description": "Livro não existe"
                    }
                }
            }

    },
    "/orders/{user}":{
        "get": {
            "tags": ["Compras user"],
            "description": "Obtem as compras efetuadas por um user",
            "produces":["application/json"],
            "parameters": [
                {
                    "in": "path",
                    "name": "user",
                    "required": true
                }
            ],
            "responses": {
                "200":{
                    "description": "Cliente existe"
                },
                "404":{
                    "description": "Not found"
                }
            }
        }
    },
    "/editarbook/{id}":{
        "post":{
            "tags":["Editar livro"],
            "description": "Permite editar livro",
            "produces":["application/json"],
            "consumes":["application/x-www-form-urlencoded"],
            "schemes": ["http"
            ],
            "parameters": [
                {
                    "schema": {
                        "$ref": "#/definitions/book"
                    }
                }
            ],
            "responses": {
                "200":{
                    "description": "Livro editado"
                }
            }
        }
    }
    
    
  

},
"definitions": {
    "book":{
        "type":"object",
        "required": [
            "title",
            "image",
            "description",
            "price",
            "stock",
            "category"
        ],
    "properties": {
    "title": {
        "type": "string"
    },
    "image": {
        "type": "string"
    },
    "description": {
        "type": "string"
    },
    "price": {
        "type": "number"
    },

    "stock":{
        "type": "number"
    },
    "category":{
        "type":"string"
    }
}
    }
}
}